Move SilentPaymentAddress and Network to utils
These structs are not meant to be used for the default sending and
receiving flow, but can still be quite useful. Therefore the utils
module is probably the best place for them.
