+++ src/common.rs
@@ -2,7 +2,6 @@ use crate::utils::hash::SharedSecretHash;
 use crate::Result;
 use bitcoin_hashes::Hash;
 use secp256k1::{PublicKey, Scalar, Secp256k1, SecretKey};
-use serde::{Deserialize, Serialize};
 
 pub(crate) fn calculate_t_n(ecdh_shared_secret: &PublicKey, k: u32) -> Result<SecretKey> {
     let hash = SharedSecretHash::from_ecdh_and_k(ecdh_shared_secret, k).to_byte_array();
@@ -17,11 +16,4 @@ pub(crate) fn calculate_P_n(B_spend: &PublicKey, t_n: Scalar) -> Result<PublicKe
     let P_n = B_spend.add_exp_tweak(&secp, &t_n)?;
 
     Ok(P_n)
-}
-
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, Deserialize, Serialize)]
-pub enum Network {
-    Mainnet,
-    Testnet,
-    Regtest,
 }
